move loaded_image out of its own module
